Improve decision handling loop and logging
